backend/.ipynb_checkpoints/requirements-checkpoint.txt
langchain-community
langchain-google-vertexai
pypdf
chromadb
langchain-chroma